# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "API Heroes Node JS",
    "contact": {
      "name": "Brian Robert",
      "email": "brian.robert16@hotmail.com"
    },
    "license": {
      "name": "MIT License"
    }
  },
  "servers": [{
    "url": "http://localhost:4000/",
    "description": "Local server"
  }],
  "security": [{
    "ApiKeyAuth": [

    ]
  }],
  "paths": {
    "/heroes": {
      "get": {
        "tags": [
          "CRUD Herois"
        ],
        "description": "Get Herois",
        "operationId": "getHerois",
        "responses": {
          "200": {
            "description": "Herois",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Herois"
                }
              }
            }
          },
          "400": {
            "description": "Missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "companyId is missing",
                  "internal_code": "missing_parameters"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CRUD Herois"
        ],
        "description": "Cadastro Herois",
        "operationId": "postHeroisId",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeroiPost"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Herois",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeroiPostReturn"
                }
              }
            }
          },
          "400": {
            "description": "Missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "companyId is missing",
                  "internal_code": "missing_parameters"
                }
              }
            }
          }
        }
      }
    },
    "/heroes/{id}": {
      "get": {
        "tags": [
          "CRUD Herois"
        ],
        "description": "Get Herois por Id",
        "operationId": "getHeroisId",
        "parameters": [{
          "name": "id",
          "in": "path",
          "schema": {
            "type": "integer"
          },
          "required": true
        }],
        "responses": {
          "200": {
            "description": "Herois",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Heroi"
                }
              }
            }
          },
          "400": {
            "description": "Missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "companyId is missing",
                  "internal_code": "missing_parameters"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "CRUD Herois"
        ],
        "description": "Delete Herois por Id",
        "operationId": "deleteHeroisId",
        "parameters": [{
          "name": "id",
          "in": "path",
          "schema": {
            "type": "integer"
          },
          "required": true
        }],
        "responses": {
          "200": {
            "description": "Herois"
          }
        }
      },
      "put": {
        "tags": [
          "CRUD Herois"
        ],
        "description": "Update Herois por Id",
        "operationId": "updateHeroisId",
        "parameters": [{
          "name": "id",
          "in": "path",
          "schema": {
            "type": "integer"
          },
          "required": true
        }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeroiPost"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Herois",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeroiPostReturn"
                }
              }
            }
          },
          "400": {
            "description": "Missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "companyId is missing",
                  "internal_code": "missing_parameters"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Heroi": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "description": "User identification number",
            "example": 1234
          },
          "identidade_secreta": {
            "type": "string",
            "example": "Nome Verdadeiro"
          },
          "nome_heroi": {
            "type": "string",
            "example": "Nome Fantasia"
          },
          "poder": {
            "type": "string",
            "example": "Poder Heroi"
          }
        }
      },
      "HeroiPost": {
        "type": "object",
        "properties": {
          "identidade_secreta": {
            "type": "string",
            "example": "Nome Verdadeiro"
          },
          "nome_heroi": {
            "type": "string",
            "example": "Nome Fantasia"
          },
          "poder": {
            "type": "string",
            "example": "Poder Heroi"
          },
          "imagem": {
            "type": "string"
          }
        }
      },
      "HeroiPostReturn": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "identidade_secreta": {
            "type": "string",
            "example": "Nome Verdadeiro"
          },
          "nome_heroi": {
            "type": "string",
            "example": "Nome Fantasia"
          },
          "poder": {
            "type": "string",
            "example": "Poder Heroi"
          },
          "message": {
            "type": "string",
            "example": "Registro Cadastrado"
          }
        }
      },
      "Herois": {
        "type": "object",
        "properties": {
          "herois": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Heroi"
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "internal_code": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  }
}